+++ packages/javascript/lib/node/crypto/md5Hash.ts
@@ -1,5 +1,8 @@
 import { createHash } from 'crypto';
+import type { BinaryToTextEncoding } from 'node:crypto';
 
-export function md5Hash(str: string): string {
-	return createHash('md5').update(str).digest('hex');
+export function md5Hash(str: string, encoding?: BinaryToTextEncoding): string {
+	return createHash('md5')
+		.update(str)
+		.digest(encoding ?? 'hex');
 }