+++ packages/javascript/lib/node/crypto/sha1Hash.ts
@@ -1,5 +1,8 @@
 import { createHash } from 'crypto';
+import type { BinaryToTextEncoding } from 'node:crypto';
 
-export function sha1Hash(str: string): string {
-	return createHash('sha1').update(str).digest('hex');
+export function sha1Hash(str: string, encoding?: BinaryToTextEncoding): string {
+	return createHash('sha1')
+		.update(str)
+		.digest(encoding ?? 'hex');
 }